Convert all of bfast to f64.
REPORT:

np.all(breaks_o == breaks_p): PASSED
np.all(hist_o == hist_p): PASSED
np.all(np.isclose(means_o, means_p)): FAILED
| Num. differences 207
| Expected [nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan
 nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan
 nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan
 nan nan  0. nan  0.  0.  0.  0. nan nan  0.  0. nan nan nan nan nan nan
 nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan
 nan nan nan nan nan nan nan nan nan nan nan nan nan nan  0. nan  0. nan
 nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan  0. nan  0.
  0. nan nan nan nan nan  0.  0.  0. nan  0. nan nan  0.  0.  0.  0.  0.
 nan nan nan nan nan nan  0.  0. nan nan nan  0.  0.  0.  0.  0.  0. nan
  0.  0.  0.  0.  0.  0.  0. nan nan nan nan nan nan nan nan nan nan nan
 nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan
 nan nan nan nan nan nan nan  0. nan]
| Actual   [nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan
 nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan
 nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan
 nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan
 nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan
 nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan
 nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan
 nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan
 nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan
 nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan
 nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan nan
 nan nan nan nan nan nan nan nan nan]
| Relative absolute error
| Max error  5.94299e-08 (0.0000%)
| Min error  3.85136e-14 (0.0000%)
| Mean error 2.14495e-08 (0.0000%)
np.all(np.isclose(magnitudes_o, magnitudes_p)): FAILED
| Num. differences 116
| Expected [...]
| Actual   [...]
| Relative absolute error
| Max error         inf (inf%)
| Min error  3.37950e-13 (0.0000%)
| Mean error        inf (inf%)
np.all(valids_o == valids_p): PASSED

NOTES:

Much better accuracy with f64. Magnitudes are still way off,
but note that only 116 out of 100k are.

+++ bfast/monitor/opencl/base.py
@@ -304,7 +304,7 @@ class BFASTMonitorOpenCL(BFASTMonitorBase):
 
     def _fit_single_preprocess(self, data, dates, nan_value):
         start = time.time()
-        mapped_indices = map_indices(dates).astype(numpy.int32)
+        mapped_indices = map_indices(dates).astype(numpy.int64)
         N = data.shape[0]
         self.n = compute_end_history(dates, self.start_monitor)
         self.lam = compute_lam(N, self.hfrac, self.level, self.period)
